add Lazy loading to ef core

+++ src/Restaurants.Application/Commands/Categories/DeleteCategory/DeleteCategoryCommandHandler.cs
@@ -34,12 +34,15 @@ public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryComman
                 throw new UnAuthorizedException("You are not authorized to delete category from this restaurant.");
 
             var category = await _categoriesRepository.GetByIdAsync(request.Id)
-                ?? throw new ResourseNotFoundException(nameof(Category), request.Id.ToString());
+                ?? throw new ResourseNotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
 
-            _logger.LogInformation("Deleting Category: {@Category}", category);
+            if(category.RestaurantId != request.RestaurantId)
+                throw new UnAuthorizedException("Category does not belong to this restaurant.");
+
+            _logger.LogInformation("Deleting Category with ID: {CategoryId}", category.Id);
 
             await _categoriesRepository.DeleteAsync(category);
-            _logger.LogInformation("Category: {@Category} deleted successfully.", category);
+            _logger.LogInformation("Category with ID: {CategoryId} deleted successfully.", category.Id);
         }
         catch(UnAuthorizedException ex)
         {
